mpfr: Add vasprintf patch from upstream
The formatted output functions (mpfr_*printf) may output/return an
incorrect string when the value is rounded up to the next power of 10
(for instance, 999.99999... printed with "%.20RNg" may give "100"
instead of "1000"). This bug, which has always been present (since these
functions were introduced in MPFR 2.4.0), is fixed by the vasprintf
patch (which also provides testcases). Note: MPFR 3.1.0 also introduced
a regression in some other cases, such as -9.996 in 32 bits printed with
"%.2Rg".

Signed-off-by: Bernhard Rosenkraenzer <Bernhard.Rosenkranzer@linaro.org>