mpfr: Add gmp4lcompat patch from upstream
The testsuite (make check) fails when MPFR is built against GMP 4.1.*
(which is a very old version). This is not a bug in the MPFR library,
but in the tests, which do not take into account the fact that the
default random number generator of GMP changed between versions 4.1 and
4.2. This problem is fixed by the gmp41compat patch, which also updates
the MPFR manual to mention this change (in the API Compatibility
chapter).

Signed-off-by: Bernhard Rosenkraenzer <Bernhard.Rosenkranzer@linaro.org>