mpfr: Add logging-freeze patch from upstream
When logging is used (note that logging support must have explicitly
been enabled in the MPFR build), e.g. with MPFR_LOG_ALL=1, the program
may randomly freeze (observed with dynamic linking), take a lot of
memory or have other random behavior. The problem actually occurs in the
following functions: mpfr_add_d, mpfr_add_ui and mpfr_mul_d; but
mpfr_add_ui is used by other functions (directly or indirectly). This is
due to a typo in an argument for the internal logging macro
MPFR_LOG_FUNC: instead of the variable mpfr_log_prec (whose value is 6
by default), the function pointer mpfr_get_prec is used, and it has an
arbitrary value. This bug (#13933) is fixed by the logging-freeze patch.

Signed-off-by: Bernhard Rosenkraenzer <Bernhard.Rosenkranzer@linaro.org>