mpfr: Add large-prec patch from upstream
Since the change of mpfr_prec_t to a signed type, it is no longer
possible to use precisions close to MPFR_PREC_MAX in some cases where it
used to work. This problem is fixed by the large-prec patch, which also
provides a simple testcase. Note that:

-   Only machines with a 32-bit precision type are affected by this
problem in practice ("32-bit machines", or "64-bit machines" where MPFR
is built with a 32-bit ABI, e.g. with the -m32 compiler option).
-   In general, intermediate variables are used and the precision is
increased internally, so that MPFR may abort in very large precisions,
even less than MPFR_PREC_MAX, because of this type limitation or a lack
of memory.
-   In very large precisions, there may still be integer overflows (due
to increased working precision) instead of a clean abort.

Signed-off-by: Bernhard Rosenkraenzer <Bernhard.Rosenkranzer@linaro.org>